refactor(platform-core): 领域层按 features 拆分到独立文件
把 domains.ts 拆分为可独立维护的文件：context.ts（ConnectionContext /
PlatformDomainService）、features/{connection,pull-request,comment,media}.ts（各领域
接口 + 基类）、adapter.ts（容器 + composePlatformAdapter）。index.ts 对外导出不变，
纯文件组织调整，避免单文件随基类共享逻辑增长而膨胀。
