+++ packages/platform-core/src/index.ts
@@ -16,21 +16,15 @@ export {
   extractApiMessage,
   resolveConnectionFetch,
 } from './http.js';
-export type {
-  PlatformConnection,
-  PullRequestService,
-  CommentService,
-  MediaService,
-  PlatformAdapter,
-  ConnectionContext,
-  PlatformServices,
-} from './domains.js';
-export {
-  MutableConnectionContext,
-  PlatformDomainService,
-  BaseConnection,
-  BasePullRequestService,
-  BaseCommentService,
-  BaseMediaService,
-  composePlatformAdapter,
-} from './domains.js';
+export type { ConnectionContext } from './context.js';
+export { MutableConnectionContext, PlatformDomainService } from './context.js';
+export type { PlatformConnection } from './features/connection.js';
+export { BaseConnection } from './features/connection.js';
+export type { PullRequestService } from './features/pull-request.js';
+export { BasePullRequestService } from './features/pull-request.js';
+export type { CommentService } from './features/comment.js';
+export { BaseCommentService } from './features/comment.js';
+export type { MediaService } from './features/media.js';
+export { BaseMediaService } from './features/media.js';
+export type { PlatformAdapter, PlatformServices } from './adapter.js';
+export { composePlatformAdapter } from './adapter.js';
